refactor(localization): relocate `loggerScopes`
