+++ src/modules/events/handlers/debug.ts
@@ -1,7 +1,7 @@
 import { Events } from "discord.js";
 import AresEventHandler from "../aresEventHandler";
 import logger from "../../logger/logger";
-import { LoggerScopes } from "../../../util/loggerScopes";
+import { LoggerScopes } from "../../logger/loggerScopes";
 
 export default new AresEventHandler(Events.Debug, false, true, (info) => {
   logger.debug(LoggerScopes.EventsHandler, { debug: info });
